Рефакторинг компонентов Tabs и MoviePage
